requirements.txt
Flask==2.0.3
torch==1.10.0
torchvision==0.11.1
joblib==1.0.1
Pillow==8.4.0
numpy==1.21.4
